+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - [TD-4604] Include actions in `POST /api/rule_implementations/search` and `POST
  /api/data_structures/search` responses
 
+ ### Fixed
+
+ - [TD-4527] Prevent showing logically deleted sources
+
 ## [4.39.0] 2022-03-07
 
 ### Added